+++ Core/VideoHud.h
@@ -13,9 +13,9 @@ private:
 	static const vector<uint32_t> _recordIcon;
 
 	void BlendColors(uint32_t* output, uint32_t input);
-	bool DisplayControllerInput(ControlDeviceState &state, int inputPort, uint8_t *outputBuffer, FrameInfo &frameInfo, OverscanDimensions &overscan, uint32_t displayIndex);
-	void DrawMovieIcons(uint8_t *outputBuffer, FrameInfo &frameInfo, OverscanDimensions &overscan);
+	bool DisplayControllerInput(ControlDeviceState &state, int inputPort, uint32_t *outputBuffer, FrameInfo &frameInfo, OverscanDimensions &overscan, uint32_t displayIndex);
+	void DrawMovieIcons(uint32_t *outputBuffer, FrameInfo &frameInfo, OverscanDimensions &overscan);
 
 public:
-	void DrawHud(uint8_t *outputBuffer, FrameInfo frameInfo, OverscanDimensions overscan);
+	void DrawHud(uint32_t *outputBuffer, FrameInfo frameInfo, OverscanDimensions overscan);
 };
